// app/src/enums/NotificationType.ts
export enum NotificationType {
    FRIEND_REQUEST_RECEIVED = "FRIEND_REQUEST_RECEIVED",
    FRIEND_REQUEST_ACCEPTED = "FRIEND_REQUEST_ACCEPTED",
    EVENT_INVITE_RECEIVED = "EVENT_INVITE_RECEIVED",
    EVENT_ATTENDANCE_UPDATE = "EVENT_ATTENDANCE_UPDATE",
    EVENT_EDITTED = "EVENT_EDITTED",
    COMMUNITY_EVENT_CREATED = "COMMUNITY_EVENT_CREATED",
    COMMUNITY_INVITE_RECEIVED = "COMMUNITY_INVITE_RECEIVED",
    COMMUNITY_INVITE_ACCEPTED = "COMMUNITY_INVITE_ACCEPTED",
}